requirements.txt
anaconda-client==1.11.2
anaconda-navigator==2.4.0
anyio==4.8.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
atomicwrites==1.4.0
attrs==23.2.0
Automat==20.2.0
autopep8==1.6.0
autoviz==0.1.905
Babel==2.11.0
backports.functools-lru-cache==1.6.4
backports.tarfile==1.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.6.0
brotlipy==0.7.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
channels-redis==4.2.0
chardet==4.0.0
charset-normalizer==3.4.0
chroma-py==0.1.0.dev1
click==8.1.7
clyent==1.2.2
colorcet==3.1.0
colour==0.1.5
comm==0.2.2
conda==23.3.1
conda-build==3.24.0
conda-content-trust==0.1.3
conda-pack==0.6.0
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
conda-repo-cli==1.0.41
conda-token==0.4.0
conda-verify==3.4.2
config==0.5.1
constantly==15.1.0
contourpy==1.3.0
cookiecutter==1.7.3
cryptography==43.0.1
cssselect==1.1.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==5.1.1
django-cors-headers==4.4.0
django-crispy-forms==2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
emoji==2.14.0
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
fonttools==4.54.1
fqdn==1.5.1
fsspec==2024.9.0
gnureadline==8.1.2
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
h11==0.14.0
HeapDict==1.0.1
holoviews==1.19.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hvplot==0.11.1
idna==3.10
imbalanced-learn==0.10.1
importlib_resources==6.4.5
iniconfig==1.1.1
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itypes==1.2.0
jedi==0.19.1
Jinja2==3.1.4
jinja2-time==0.2.0
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
linkify-it-py==2.0.3
mapboxgl==0.10.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.0
munkres==1.1.4
mysqlclient==2.2.4
navigator-updater==0.3.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.1.2
oauthlib==3.2.2
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandas-dq==1.29
pandocfilters==1.5.1
panel==1.5.2
param==2.1.1
parso==0.8.4
pathlib==1.0.1
patsy==0.5.6
pexpect==4.9.0
pillow==11.0.0
pip==22.3.1
platformdirs==4.3.6
plotly==5.9.0
pluggy==1.0.0
ply==3.11
pooch==1.4.0
poyo==0.5.0
prometheus-client==0.14.1
prompt_toolkit==3.0.48
Protego==0.1.16
proto-plus==1.24.0
protobuf==5.28.2
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyamg==5.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.10.0
pycosat==0.6.4
pycparser==2.22
pyct==0.5.0
pycurl==7.45.1
PyDispatcher==2.0.5
pydocstyle==6.3.0
pyerfa==2.0.0
pyflakes==3.0.1
pygame==2.6.1
Pygments==2.18.0
PyHamcrest==2.0.2
PyJWT==2.8.0
pylint==2.16.2
pylint-venv==2.3.0
pyls-spyder==0.4.0
pymongo==3.12.3
PyMySQL==1.1.1
pyobjc-core==9.0
pyobjc-framework-Cocoa==9.0
pyobjc-framework-CoreServices==9.0
pyobjc-framework-FSEvents==9.0
pyodbc==4.0.34
pyOpenSSL==23.0.0
pyparsing==3.2.0
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.9.0.post0
python-decouple==3.8
python-json-logger==3.2.1
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
python-slugify==5.0.2
python-snappy==0.6.1
pytoolconfig==1.2.5
pytz==2024.2
pyviz_comms==3.0.3
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==26.2.0
QDarkStyle==3.0.2
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.4.0
QtPy==2.2.0
queuelib==1.5.0
redis==5.0.3
referencing==0.33.0
regex==2024.9.11
requests==2.32.3
requests-file==1.5.1
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rope==1.7.0
rpds-py==0.18.0
rsa==4.9
Rtree==1.0.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.17.21
scikit-image==0.19.3
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.8.0
seaborn==0.13.2
Send2Trash==1.8.3
service-identity==18.1.0
setuptools==75.2.0
shortuuid==1.0.12
simplejson==3.19.2
sip==6.6.2
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.1
spyder-kernels==2.4.1
SQLAlchemy==1.4.39
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.4
sweetviz==2.3.1
sympy==1.11.1
tables==3.7.0
tabulate==0.8.10
TBB==0.2
tblib==1.7.0
tenacity==8.0.1
terminado==0.17.1
text-unidecode==1.3
textblob==0.18.0.post0
textdistance==4.2.1
threadpoolctl==3.5.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss2==1.2.1
tldextract==3.2.0
tokenizers==0.11.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
toolz==0.12.0
torch==1.12.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.24.0
tweepy==4.14.0
Twisted==22.2.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
ujson==5.4.0
Unidecode==1.2.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
w3lib==1.21.0
watchdog==2.1.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
whatthepatch==1.0.2
wheel==0.38.4
widgetsnbextension==4.0.13
wordcloud==1.9.3
wrapt==1.14.1
wurlitzer==3.0.2
xarray==2022.11.0
xgboost==1.6.2
xlrd==2.0.1
xlwings==0.29.1
xyzservices==2024.9.0
yapf==0.31.0
zict==2.1.0
zipp==3.11.0
zope.interface==5.4.0
zstandard==0.19.0
